Fix media actions in filtered playlist items list
